- Crafting fixed in 1.0.0 - Server compatible with the weapons and equip again. (Some Packets were incompatible) - fixed bucket bugs (not all) - Fixed clients getting crashed by wrong Pickups - fixed nearly all mob drops. (Check wheather they are burning is missing Big Grin) - maybe some other things I can´t recall atm Big Grin
git-svn-id: http://mc-server.googlecode.com/svn/trunk@94 0a769ca7-a7f5-676a-18bf-c427514a06d6

+++ source/cInventory.cpp
@@ -320,7 +320,7 @@ void cInventory::SendSlot( int a_SlotNum )
 	{
 		cPacket_InventorySlot InventorySlot;
 		InventorySlot.m_ItemCount = Item->m_ItemCount;
-		InventorySlot.m_ItemID = (short)Item->m_ItemID;
+		InventorySlot.m_ItemID = Item->m_ItemID;
 		InventorySlot.m_ItemUses = (char)Item->m_ItemHealth;
 		InventorySlot.m_SlotNum = (short)a_SlotNum;
 		InventorySlot.m_WindowID = 0; // Inventory window ID
